refactor booking modals
